src/producer/requirements.txt
python-dotenv==1.2.1
faker==38.0.0
jsonschema==4.25.1
confluent_kafka==2.12.2
